requirements.txt
numpy==2.2.6
opencv-contrib-python==4.11.0.86
pillow==11.2.1
pywin32==310
